+++ imaginarycss/inst/include/barry/barrayvector-meat.hpp
@@ -1,5 +1,3 @@
-#include "barrayvector-bones.hpp"
-
 #ifndef BARRY_BARRAYVECTOR_MEAT_HPP
 #define BARRY_BARRAYVECTOR_MEAT_HPP 1
 
@@ -18,7 +16,7 @@ inline void BArrayVector<Cell_Type,Data_Type>::init_vec() {
     } else {
 
         for (const auto& a : Array->el_ji[i])
-            vec.push_back(std::make_pair<uint, Cell<Cell_Type>>(a.first, *(a.second)));
+            vec.push_back(std::make_pair<size_t, Cell<Cell_Type>>(a.first, *(a.second)));
 
     }
 
@@ -38,7 +36,7 @@ inline bool BArrayVector<Cell_Type,Data_Type>::is_col() const noexcept {
 }
 
 template<typename Cell_Type, typename Data_Type>
-inline uint BArrayVector<Cell_Type,Data_Type>::size() const noexcept {
+inline size_t BArrayVector<Cell_Type,Data_Type>::size() const noexcept {
 
     if (dim == 0u)
         return Array->el_ij[i].size();
@@ -70,8 +68,8 @@ inline std::vector< Cell_Type >::const_iterator BArrayVector<Cell_Type,Data_Type
 template<typename Cell_Type,typename Data_Type>
 inline void BArrayVector<Cell_Type,Data_Type>::operator=(const Cell_Type & val) {
     
-    uint k = 0u;
-    uint N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
+    size_t k = 0u;
+    size_t N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
     
     if (dim == 0u)
     {
@@ -92,8 +90,8 @@ inline void BArrayVector<Cell_Type,Data_Type>::operator=(const Cell_Type & val) 
 template<typename Cell_Type,typename Data_Type>
 inline void BArrayVector<Cell_Type,Data_Type>::operator+=(const Cell_Type & val) {
     
-    uint k = 0u;
-    uint N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
+    size_t k = 0u;
+    size_t N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
     
     if (dim == 0u)
     {
@@ -113,8 +111,8 @@ inline void BArrayVector<Cell_Type,Data_Type>::operator+=(const Cell_Type & val)
 template<typename Cell_Type,typename Data_Type>
 inline void BArrayVector<Cell_Type,Data_Type>::operator-=(const Cell_Type & val) {
     
-    uint k = 0u;
-    uint N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
+    size_t k = 0u;
+    size_t N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
     
     if (dim == 0u)
     {
@@ -134,8 +132,8 @@ inline void BArrayVector<Cell_Type,Data_Type>::operator-=(const Cell_Type & val)
 template<typename Cell_Type,typename Data_Type>
 inline void BArrayVector<Cell_Type,Data_Type>::operator*=(const Cell_Type & val) {
     
-    uint k = 0u;
-    uint N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
+    size_t k = 0u;
+    size_t N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
     
     if (dim == 0u)
     {
@@ -155,8 +153,8 @@ inline void BArrayVector<Cell_Type,Data_Type>::operator*=(const Cell_Type & val)
 template<typename Cell_Type,typename Data_Type>
 inline void BArrayVector<Cell_Type,Data_Type>::operator/=(const Cell_Type & val) {
     
-    uint k = 0u;
-    uint N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
+    size_t k = 0u;
+    size_t N_ = (dim == 0u) ? Array->nrow() : Array->ncol();
     
     if (dim == 0u)
     {
@@ -188,7 +186,7 @@ inline bool BArrayVector<Cell_Type,Data_Type>::operator==(const Cell_Type & val)
 
     if (dim == 0u)
     {
-        for (uint j = 0u; j < Array->ncol(); ++j)
+        for (size_t j = 0u; j < Array->ncol(); ++j)
         {
             if (Array(i, j) != val)
                 return false;
@@ -197,7 +195,7 @@ inline bool BArrayVector<Cell_Type,Data_Type>::operator==(const Cell_Type & val)
 
     } else {
 
-        for (uint j = 0u; j < Array->nrow(); ++j)
+        for (size_t j = 0u; j < Array->nrow(); ++j)
         {
             if (Array(j, i) != val)
                 return false;
@@ -225,7 +223,7 @@ inline bool BArrayVector_const<Cell_Type,Data_Type>::operator==(const Cell_Type 
     
     if (dim == 0u)
     {
-        for (uint j = 0u; j < Array->ncol(); ++j)
+        for (size_t j = 0u; j < Array->ncol(); ++j)
         {
             if (Array(i, j) != val)
                 return false;
@@ -234,7 +232,7 @@ inline bool BArrayVector_const<Cell_Type,Data_Type>::operator==(const Cell_Type 
 
     } else {
 
-        for (uint j = 0u; j < Array->nrow(); ++j)
+        for (size_t j = 0u; j < Array->nrow(); ++j)
         {
             if (Array(j, i) != val)
                 return false;
@@ -257,7 +255,7 @@ inline bool BArrayVector_const<Cell_Type,Data_Type>::operator<(const Cell_Type &
     
     if (dim == 0u)
     {
-        for (uint j = 0u; j < Array->ncol(); ++j)
+        for (size_t j = 0u; j < Array->ncol(); ++j)
         {
             if (Array(i, j) >= val)
                 return false;
@@ -266,7 +264,7 @@ inline bool BArrayVector_const<Cell_Type,Data_Type>::operator<(const Cell_Type &
 
     } else {
 
-        for (uint j = 0u; j < Array->nrow(); ++j)
+        for (size_t j = 0u; j < Array->nrow(); ++j)
         {
             if (Array(j, i) >= val)
                 return false;
@@ -284,7 +282,7 @@ inline bool BArrayVector_const<Cell_Type,Data_Type>::operator<=(const Cell_Type 
     
     if (dim == 0u)
     {
-        for (uint j = 0u; j < Array->ncol(); ++j)
+        for (size_t j = 0u; j < Array->ncol(); ++j)
         {
             if (Array(i, j) > val)
                 return false;
@@ -293,7 +291,7 @@ inline bool BArrayVector_const<Cell_Type,Data_Type>::operator<=(const Cell_Type 
 
     } else {
 
-        for (uint j = 0u; j < Array->nrow(); ++j)
+        for (size_t j = 0u; j < Array->nrow(); ++j)
         {
             if (Array(j, i) > val)
                 return false;
